Fix asyncio for failing test
